+++ NomeroffNet/mcm/__init__.py
@@ -1,0 +1,1 @@
+from .mcm import *