+++ NomeroffNet/mcm/mcm.py
@@ -1,16 +1,15 @@
-import json
 import os
 from tensorflow.python.client import device_lib
 import urllib.request
 from tqdm import tqdm
 import pathlib
+from .latest import latest_models
 
 # load latest paths
 dirpath = os.getcwd()
-with open(os.path.join(dirpath, "../models/latest.json")) as jLatest:
-    latest_models = json.load(jLatest)
 
-def load_last_models():
+
+def show_last_models():
     print(latest_models)
 
 def get_mode():
@@ -22,7 +21,21 @@ def get_mode():
 
 device_mode = get_mode()
 
+def ls():
+    models_list = []
+    for r, d, f in os.walk(os.path.join(os.path.dirname(os.path.realpath(__file__)), "./models")):
+        for file in f:
+            models_list.append(file)
+    return models_list
 
+def rm(model_name):
+    models_list = []
+    for r, d, f in os.walk(os.path.join(os.path.dirname(os.path.realpath(__file__)), "./models")):
+        for file in f:
+            if file == model_name:
+                os.remove(os.path.join(r, file))
+                return True
+    return False
 
 class DownloadProgressBar(tqdm):
     def update_to(self, b=1, bsize=1, tsize=None):
@@ -38,7 +51,7 @@ def download_url(url, output_path):
 
 def download_latest_model(detector, model_name, ext="h5", mode = device_mode):
     info = latest_models[detector][model_name][ext]
-    info["path"] = os.path.join(dirpath, "../models", detector, model_name, os.path.basename(info[mode]))
+    info["path"] = os.path.join(os.path.dirname(os.path.realpath(__file__)), "./models", detector, model_name, os.path.basename(info[mode]))
 
     p = pathlib.Path(os.path.dirname(info["path"]))
     p.mkdir(parents=True, exist_ok=True)